refactor: Extract common code into lib directory
