+++ lib/database.py
@@ -1,0 +1,7 @@
+import os
+import psycopg2
+from datetime import datetime
+
+def get_db_connection():
+    """Get a PostgreSQL database connection"""
+    return psycopg2.connect(os.getenv('DATABASE_URL'))
