+++ lib/types.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+class ChannelMode(Enum):
+    RESTRICTED = "RESTRICTED"
+    FREE = "FREE"
